requirements.txt
Flask==1.1.4
stripe==2.56.0
python-dotenv==0.19.0